requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dataclasses==0.8
DateTime==4.4
Flask==2.0.3
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
PyAudio==0.2.11
pymongo==3.2
pypinyin==0.46.0
pytz==2022.1
qrcode==7.3.1
requests==2.27.1
selenium==3.141.0
typing-extensions==4.1.1
urllib3==1.26.9
Wave==0.0.2
Werkzeug==2.0.3
ybc-exception==1.2.1
ybc-pinyin==1.0.3
ybc-speech1==1.0.3
zipp==3.6.0
zope.interface==5.4.0
